refactor(emul-23): moved Cta into a subfolder
